# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "ACROSS Server",
        "summary": "Astrophysics Cross-Observatory Science Support (ACROSS)",
        "description": "Server providing tools and utilities for various NASA missions to aid in coordination of large observation efforts.",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/user/": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get Many",
                "operationId": "get_users",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/across_server__routes__user__schemas__User"
                                    },
                                    "type": "array",
                                    "title": "Response Get Users"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The user does not exist."
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create a user",
                "description": "Create a new user for ACROSS.",
                "operationId": "create_user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "The user does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/user/{user_id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Read a user",
                "description": "Read a user by a user ID.",
                "operationId": "get_user",
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "UUID of the user"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return a user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/across_server__routes__user__schemas__User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The user does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "Update a user",
                "description": "Update a user's information for the allowed properties.",
                "operationId": "update_user",
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "UUID of the user"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The updated user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/across_server__routes__user__schemas__User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The user does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete a user",
                "description": "Permanently delete a user from the ACROSS system. This will also delete all associated relations for the user.",
                "operationId": "delete_user",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "User Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Upon successful deletion, the user is returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/across_server__routes__user__schemas__User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The user does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{user_id}/invite": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Read a user's group invites",
                "description": "Read a user's group invites by a user ID.",
                "operationId": "get_group_invites",
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "UUID of the user"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of group invites for a user id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/across_server__routes__user__schemas__GroupInvite"
                                    },
                                    "title": "Response 200 Get Group Invites"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The user does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{user_id}/invite/{invite_id}": {
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "Accept a group invitation",
                "description": "Accept a group invitation to join a user group.",
                "operationId": "accept_invite",
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "UUID of the user"
                        }
                    },
                    {
                        "name": "invite_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Invite Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The group invitation has been accepted"
                    },
                    "404": {
                        "description": "The user does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Decline a group invitation",
                "description": "Decline a group invitation to refuse joining a user group.",
                "operationId": "decline_invite",
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "UUID of the user"
                        }
                    },
                    {
                        "name": "invite_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Invite Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The group invitation has been declined"
                    },
                    "404": {
                        "description": "The user does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{user_id}/group/{group_id}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Leave a group",
                "description": "Leave a group by id",
                "operationId": "leave_group",
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "UUID of the user"
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Group Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully left the group"
                    },
                    "404": {
                        "description": "The user does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/group/{group_id}/user/{user_id}": {
            "delete": {
                "tags": [
                    "Group"
                ],
                "summary": "Remove a user from a group",
                "description": "Remove a user from a group by ID",
                "operationId": "remove_user",
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "User Id"
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "UUID of the group"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "404": {
                        "description": "The group does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/group/{group_id}/invite": {
            "get": {
                "tags": [
                    "Group Invite"
                ],
                "summary": "Read group invites",
                "description": "Read all group invites for a group",
                "operationId": "get_invites",
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "UUID of the group"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of group invites",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/across_server__routes__group__invite__schemas__GroupInvite"
                                    },
                                    "title": "Response 200 Get Invites"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The group or invite does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Group Invite"
                ],
                "summary": "Create and send a group invite",
                "description": "Create and send a group invite to a user",
                "operationId": "send_invite",
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "UUID of the group"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GroupInviteCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "The created group invite",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/across_server__routes__group__invite__schemas__GroupInvite"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The group or invite does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "UserCreate": {
                "properties": {
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "first_name": {
                        "type": "string",
                        "title": "First Name"
                    },
                    "last_name": {
                        "type": "string",
                        "title": "Last Name"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    }
                },
                "type": "object",
                "required": [
                    "username",
                    "first_name",
                    "last_name",
                    "email"
                ],
                "title": "UserCreate"
            },
...
            "across_server__routes__user__schemas__GroupInvite": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "group": {
                        "$ref": "#/components/schemas/GroupInviteGroupDetails"
                    },
                    "sender": {
                        "$ref": "#/components/schemas/UserInfo"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "group",
                    "sender"
                ],
                "title": "GroupInvite"
            },
...
            "across_server__routes__user__schemas__User": {
                "properties": {
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "first_name": {
                        "type": "string",
                        "title": "First Name"
                    },
                    "last_name": {
                        "type": "string",
                        "title": "Last Name"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "groups": {
                        "items": {
                            "$ref": "#/components/schemas/across_server__routes__user__schemas__Group"
                        },
                        "type": "array",
                        "title": "Groups"
                    },
                    "roles": {
                        "items": {
                            "$ref": "#/components/schemas/RoleBase"
                        },
                        "type": "array",
                        "title": "Roles"
                    },
                    "group_roles": {
                        "items": {
                            "$ref": "#/components/schemas/across_server__routes__user__schemas__GroupRole"
                        },
                        "type": "array",
                        "title": "Group Roles"
                    },
                    "received_invites": {
                        "items": {
                            "$ref": "#/components/schemas/across_server__routes__user__schemas__GroupInvite"
                        },
                        "type": "array",
                        "title": "Received Invites"
                    }
                },
                "type": "object",
                "required": [
                    "username",
                    "first_name",
                    "last_name",
                    "email",
                    "id",
                    "groups",
                    "roles",
                    "group_roles",
                    "received_invites"
                ],
                "title": "User"
            },
...
        },
        "securitySchemes": {
            "Authorization": {
                "type": "http",
                "description": "Enter your access token.",
                "scheme": "bearer"
            }
        }
    }
}